requirements.txt
PyQt6==6.6.0
PyQt6-WebEngine==6.6.0
pyqtgraph==0.13.3
pyserial==3.5
